REVAMP
Changed the way shit works. clicking the buttons no longer removes the ships from the board. You can still place and replace the ships to your heart's content, you just can' completely remove a ship from the board once it's been placed. Replacing this feature is the fact that you can place a ship on coordinates it used to occupy/partially occupy. I couldn't figure out how to do this before, so I made it that the player could remove ships in order to place them this way. Also you may notice that this version is OVER 100 LINES SHORTER lets go dude I implemented a lot of the really bulky ifs and switches as way more compact shade() and check() methods that can be used whenever. shade() shades coordinates to colors of your choice, and check() checks if it's safe to place a ship down. Also I got rid of readyCheck() because there's no longer an instance where you'd ever stop being ready once you started, and the method was only being called once and it's only one line, so there's no reason for it to be a method. This is big progress tbh, before I was so worried about bulletproofing that I wound up with huge amounts of code and really clunky controls for the ship placement, but now it's basically  to the point where arranging your ships on the grid is as easy and fluid as a physical game of battleship, and it's way more compact