Aesthetic tweaks + error handling
changed the color constants to more "battleship-esque" colors, got rid of the angry red color when you hover your ship out of bounds or over another ship. This was fine but I felt that it was sending inconsistent signals, since it would go red when you hovered a ship over itself (which would be a legal ship placement). So yeah the try-catches for mouseEntered() and mouseExited() are empty now as there's nothing to put there anymore, they just exist to stop it from spitting out stack traces when you hover your ship out of bounds